refactor: rename backend restart recovery test to SpringContextRestartRecoveryIntegrationTest and update references
