+++ src/test/kotlin/org/machikoro/server/service/BackendContainerRestartRecoveryTest.kt
@@ -1,3 +1,3 @@
 package org.machikoro.server.service
 
-// Replaced by BackendRestartRecoveryIntegrationTest.kt
+// Replaced by SpringContextRestartRecoveryIntegrationTest.kt
